feat: migrate swap into AlertLogger
