+++ p2p_wallet/Services/Logger/LogManagerLoggers/AlertLogger.swift
@@ -1,14 +1,9 @@
 import Foundation
 
-struct AlertLoggerError<T: Codable>: Error, Codable {
-    var title: String
-    var message: T
-}
-
 final class AlertLogger: LogManagerLogger {
     var supportedLogLevels: [LogLevel] = [.alert]
     
-    private let url = URL(string: "https://oncall.keyapp.org/integrations/v1/formatted_webhook/yQ9zMIbgg64nhdKC1TAViG53t/")!
+    private let url = URL(string: .secretConfig("SWAP_ERROR_LOGGER_ENDPOINT")!)!
 
     func log(event: String, logLevel: LogLevel, data: String?) {
         Task {
@@ -16,9 +11,17 @@ final class AlertLogger: LogManagerLogger {
             var urlRequest = URLRequest(url: url)
             urlRequest.httpMethod = "POST"
             urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
-            urlRequest.httpBody = data?.data(using: .utf8) ?? Data()
-//            _ = try? await URLSession.shared.data(from: urlRequest)
+            let body = AlertLoggerError(title: event, message: data ?? "")
+            urlRequest.httpBody = try JSONEncoder().encode(body)
+            _ = try? await URLSession.shared.data(from: urlRequest)
         }
     }
 
+}
+
+// MARK: - Models
+
+private struct AlertLoggerError: Error, Codable {
+    var title: String
+    var message: String
 }
