--- p2p_wallet/Services/SwapErrorLogger/SwapErrorLoggerError.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public enum SwapErrorLoggerError: Error {
-    case invalidStatusCode
-}
